Use GoPerf wrappers
