+++ cmd/cpi-count/main.go
@@ -1,13 +1,9 @@
 package main
 
 import (
-	"crypto/sha256"
-	"encoding/base64"
 	"log"
 	"os"
 	"runtime"
-
-	"github.com/elastic/go-perf"
 )
 
 func init() {
@@ -21,59 +17,26 @@ func init() {
 }
 
 func main() {
-	pid := os.Getpid()
-	log.Printf("Current PID: %d\n", pid)
-
-	perfCmd := NewPerfCmd(pid)
+	perfCmd := NewPerfCmd(os.Getpid())
 	if err := perfCmd.Start(); err != nil {
 		log.Fatalf("Failed to execute perf cmd: %v\n", err)
 	}
-	log.Printf("Started perf cmd\n")
 
-	g := perf.Group{
-		CountFormat: perf.CountFormat{
-			Running: true,
-		},
+	goperf := NewGoPerf()
+	if err := goperf.StartWorkloadMeasurement(); err != nil {
+		log.Fatalf("Failed to start goperf measurement: %v\n", err)
 	}
-	g.Add(perf.Instructions, perf.CPUCycles)
 
-	p, err := g.Open(perf.CallingThread, perf.AnyCPU)
+	goperfOutput, err := goperf.End()
 	if err != nil {
-		log.Fatalf("Failed to open perf events: %s\n", err)
+		log.Fatalf("Failed to end goperf measurement: %v\n", err)
 	}
 
-	var workloadOutput string
-	gc, err := p.MeasureGroup(func() {
-		workloadOutput = heavyWorkload()
-	})
-
-	if err != nil {
-		log.Fatalf("Failed to measure perf group: %s\n", err)
-	}
-
-	p.Close()
 	perfOutput, err := perfCmd.End()
 	if err != nil {
 		log.Fatalf("Failed to end perf cmd: %v\n", err)
 	}
 
-	cycles, instrs := gc.Values[1].Value, gc.Values[0].Value
-	log.Printf("Output is %s\n", workloadOutput)
-	log.Printf("Ran for %dms\n", gc.Running.Milliseconds())
-	log.Printf("GoPerf Cycles: %d, GoPerf Instrs: %d, GoPerf CPI: %f\n", cycles, instrs, float64(cycles)/float64(instrs))
+	log.Printf("GoPerf Cycles: %d, GoPerf Instrs: %d, GoPerf CPI: %f\n", int64(goperfOutput.Cycles), int64(goperfOutput.Instrs), goperfOutput.Cycles/goperfOutput.Instrs)
 	log.Printf("PerfCmd Cycles: %d, PerfCmd Instrs: %d, PerfCmd CPI: %f\n", int64(perfOutput.Cycles), int64(perfOutput.Instrs), perfOutput.Cycles/perfOutput.Instrs)
-}
-
-func heavyWorkload() string {
-	seedStr := "1sAMsDJGtS3zNrK6MfeysFvUYOzlHqtj"
-
-	var hash string
-	hashBytes := sha256.Sum256([]byte(seedStr))
-
-	for i := 0; i < 999999; i++ {
-		hash = base64.StdEncoding.EncodeToString(hashBytes[:])
-		hashBytes = sha256.Sum256([]byte(hash))
-	}
-
-	return hash
 }
